MP-025 Диструктуризация проекта, вывод компонета элементов списка товаров и магазинов выше
