--- src/components/content/goods/goods-list/goods-list-item/goods-list-item.module.scss
@@ -1,61 +1,0 @@
-.good {
-  width: 15%;
-  min-width: 200px;
-  background-color: #eff3f6;
-  border-radius: 4px;
-  box-sizing: border-box;
-  cursor: pointer;
-  padding: 15px;
-  transition: background-color .2s ease-out;
-
-  margin-bottom: 24px;
-  margin-right: 34px;
-
-  &:hover {
-    background-color: darken(#eff3f6, 10%);
-    cursor: pointer;
-  }
-
-  color: black;
-  text-decoration: none;
-}
-
-.image {
-  width: 100%;
-  height: 20vh;
-  border-radius: 10px;
-  border: 2px solid #fff;
-  max-height: 200px;
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 10px;
-  }
-}
-
-.title {
-  text-align: center;
-  padding: 10px 0;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.price {
-  padding: 5px;
-
-  font-size: 24px;
-  font-weight: 500;
-}
-
-.addCart {
-  display: flex;
-  justify-content: center;
-  margin-top: 20px;
-
-  &:hover {
-    cursor: pointer;
-  }
-}
