--- src/components/content/markets/markets-list/markets-list-item/markets-list-item.module.scss
@@ -1,67 +1,0 @@
-.market {
-  width: calc(23% - 18px);
-  background-color: #eff3f6;
-  border-radius: 4px;
-  box-sizing: border-box;
-  cursor: pointer;
-  padding: 24px;
-  transition: background-color .2s ease-out;
-
-  margin-bottom: 24px;
-  margin-right: 24px;
-
-  &:hover {
-    background-color: darken(#eff3f6, 10%);
-    cursor: pointer;
-  }
-
-  color: black;
-  text-decoration: none;
-}
-
-.titleRow {
-  display: flex;
-  height: 64px;
-  padding-bottom: 16px;
-
-  .photo {
-    align-items: center;
-    display: flex;
-    height: 100%;
-    justify-content: center;
-    position: relative;
-
-
-
-    img {
-      border: 2px solid #fff;
-      border-radius: 50%;
-      box-sizing: border-box;
-      height: 64px;
-      width: 64px;
-    }
-  }
-
-  .titleBlock {
-    padding-left: 10px;
-  }
-
-
-  .subTitle {
-    margin: 0;
-    -webkit-line-clamp: 2;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    font-size: 11px;
-  }
-}
-
-.title {
-  margin: 0;
-  -webkit-line-clamp: 1;
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
