+++ src/components/content/markets/markets-list/markets-list.module.scss
@@ -1,9 +1,11 @@
 .mainMarkets {
+  width: 85vw;
   display: flex;
   justify-content: center;
   flex-wrap: wrap;
 
-  margin-top: 15px;
+  margin: 15px auto;
+
   border-top: 1px solid rgba(0, 0, 0, 0.36);
   padding-top: 15px;
 }
