+++ src/components/ui-components/goods-item/goods-item.module.scss
