+++ src/components/ui-components/goods-item/to-favorites-button/to-favorites-button.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ToFavoritesButton = () => {
+    return (
+        <div>
+
+        </div>
+    );
+};
+
+export default ToFavoritesButton;
