+++ src/components/ui-components/markets-item/markets-item.module.scss
